// generated/api/gameDeliveryType.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * juniors bootcamp backend 🔥
 * Апи для выполнения индивидуальных заданий
 * OpenAPI spec version: 1.0
 */

/**
 * Способ получения
 */
export type GameDeliveryType =
  (typeof GameDeliveryType)[keyof typeof GameDeliveryType];

export const GameDeliveryType = {
  steam_key: "steam_key",
  steam_gift: "steam_gift",
  epic_key: "epic_key",
  nintendo_key: "nintendo_key",
  xbox_key: "xbox_key",
  playstation_key: "playstation_key",
} as const;